Maquetación de espacio para mostrar información, obtención de dato, presenta error para imprimirlo en pantalla aunque en consola si se muetra
