+++ index.html
@@ -3,9 +3,27 @@
   <head>
     <meta charset="utf-8">
     <title>Promises,Node.js</title>
+    <link rel="stylesheet" href="static/materialize-css/dist/css/materialize.min.css">
   </head>
   <body>
     <h1>Exoplanetas</h1>
-    <script src="app.js"></script>
+    <section id="espacioExoplanetas">
+
+    </section>
+    <div class="row">
+     <div class="col s12 m6">
+       <div class="card blue-grey darken-1">
+         <div class="card-content white-text">
+           <img src="" alt="planetas">
+           <span class="card-title">Planeta Kepler</span>
+           <p>09876543321</p>
+           <p>Toda la informacion necesaria</p>
+         </div>
+       </div>
+     </div>
+    </div>
+    <script src="static/jquery/dist/jquery.min.js"></script>
+    <script src="static/materialize-css/dist/js/materialize.min.js"></script>
+    <script src="archivos/app.js"></script>
   </body>
 </html>
